2017.07.2.0
Cập nhập sửa lỗi load trang, thêm output, sửa css, sửa giao thức POST và
DELETE, giao diện mới

+++ server.php
@@ -1,21 +1,25 @@
 <?php
-function upload_file($_FILE, $rename = null) {
+function UploadFile($_FILE, $_rename = null, $_replace = false) {
     $result = false;
     $upload_dir = 'uploads';
 
     $tmp_file = $_FILE['tmp_name'];
-    if ($rename == '' || $rename == null) $rename = $_FILE['name'];
-    $destination_file = $upload_dir.'/'.$rename;
+    if ($_rename == '' || $_rename == null) $_rename = $_FILE['name'];
+    $destination_file = $upload_dir.'/'.$_rename;
 
     $upload_folder = dirname($destination_file);
     if (!is_dir($upload_folder))
         mkdir($upload_folder, 0777, true);
 
-    $upload_flag = move_uploaded_file($tmp_file, $destination_file);
+    if (file_exists($destination_file) && $_replace == false) {
+        $result = -1;
+    } else {        
+        $upload_flag = move_uploaded_file($tmp_file, $destination_file);
 
-    if ($upload_flag) {
-        $result = $destination_file;
-    }
+        if ($upload_flag) {
+            $result = $destination_file;
+        };
+    };
 
     return $result;
 }
@@ -32,9 +36,21 @@ $method = $_SERVER['REQUEST_METHOD'];
 if ($method == 'POST') {
 	if (isset($_FILES['file'])) {
         $rename = '/'.$_FILES['file']['name'];
-        $upload_file = upload_file($_FILES['file'], $rename);
+        if ($_POST['rename'] != 'false') $rename = '/'.$_POST['rename'];
 
-        if ($upload_file) {
+        $upload_file = UploadFile($_FILES['file'], $rename, $_POST['replace'] == 'true');
+
+        if ($upload_file == -1) {
+            $data = [
+                'response'  => [
+                    'success'   => false,
+                    'message'   => 'Duplicate file',
+                    'data'      => [
+                        'file'   => $rename,
+                    ],
+                ],
+            ];
+        } else if ($upload_file) {
             $data = [
                 'response'  => [
                     'success'   => true,
@@ -47,8 +63,9 @@ if ($method == 'POST') {
         }
     }
 } else if ($method == 'DELETE') {
-	if (isset($_GET['id'])) {
-        $file = $_GET['id'];
+	if (isset($_GET['id'])&&isset($_GET['ext'])) {
+        $upload_dir = 'uploads';
+        $file = $upload_dir.'/'.$_GET['id'].'.'.$_GET['ext'];
 
         if (file_exists($file)) {
             $delete_file = unlink($file);
